requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
openai==1.3.5
pandas>=2.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
# Hugging Face dependencies
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.25.0
sentencepiece>=0.1.99
huggingface-hub>=0.19.0

